% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_subplots.R
\name{determine_subplots}
\alias{determine_subplots}
\title{Find correct subplot for ranges and events}
\usage{
determine_subplots(data)
}
\arguments{
\item{data}{the data frame with ranges and events}
}
\value{
data with additional numeric column "subplot", reordered by subplot column.
}
\description{
For technical reasons (plotly), ranges and events need different subplots. We are still trying to keep events and ranges that are in the same group together.
We determine the subplot as follows: Groups appear in order of appearance in the data, events and ranges in the same group are plotted directly below each other
}
\examples{
\dontrun{
determine_subplots(data.frame(event = 1:4, start = c(Sys.Date(), Sys.Date() + 10),
                   end = c(Sys.Date(), Sys.Date() + 10), group = ""))
determine_subplots(data.frame(event = 1:4, start = c(Sys.Date(), Sys.Date() + 10),
                   end = c(Sys.Date(), Sys.Date() + 10), group = 1:2))
determine_subplots(data.frame(event = 1:3, start = c(Sys.Date(), Sys.Date() + 20,
                   Sys.Date()), end = c(Sys.Date()+10, Sys.Date()+20, Sys.Date()),
                   group = c(1,2,1)))
}
}
