% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_columns.R
\name{fix_columns}
\alias{fix_columns}
\title{Standardize column names}
\usage{
fix_columns(data, events, start, end, groups, tooltips)
}
\arguments{
\item{data}{input data frmae}

\item{events}{event column name}

\item{start}{name of start column}

\item{end}{name of end column}

\item{groups}{name of group column}

\item{tooltips}{column name of tooltips}
}
\value{
the data frame prepared for plotting
}
\description{
Standardize column names
}
\examples{
\dontrun{
fix_columns(data.frame(event = 1:4,
                       start = c("2019-01-01", "2019-01-10"),
                       end = c("2019-01-01", "2019-01-10"),
                       events = "event", start = "start", end = "end",
                       groups = "group", tooltips = "tooltip")
}
}
