% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristic_lineInterval.R
\name{heuristic_lineInterval}
\alias{heuristic_lineInterval}
\title{Heuristic for background lines in timeline graph}
\usage{
heuristic_lineInterval(data, lineInterval = NULL)
}
\arguments{
\item{data}{\code{data.frame} with (at least) columns \code{start} and \code{end}}

\item{lineInterval}{(optional) integer with suggested result (line interval
in seconds)}
}
\value{
an integer (distance in seconds to draw background lines)
}
\description{
Guesses the total range of data and calculates an appropriate background line
interval. If \code{lineInterval} is provided, checks if that is appropriate and
prints a warning message otherwise.
}
\examples{
# range is 10 days, result is 2-day-line interval
heuristic_lineInterval(data.frame(start = Sys.Date() - 10, end = Sys.Date() - 1))

}
\keyword{internal}
