% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_glued.R
\name{plot_glued}
\alias{plot_glued}
\title{PLot ranges and events together as one plot}
\usage{
plot_glued(data, plotList, title, heightsRelative)
}
\arguments{
\item{data}{the data frame to plot}

\item{plotList}{subplots as list}

\item{title}{title of the plot, can be NULL}

\item{heightsRelative}{relative heights of the subplots (sum up to 1)}
}
\value{
plotly plot
}
\description{
PLot ranges and events together as one plot
}
