% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_colors.R
\name{set_colors}
\alias{set_colors}
\title{Set column col for event colors and column fontcol for label colors}
\usage{
set_colors(data, eventcolor_column, fontcolor_column)
}
\arguments{
\item{data}{the data frame containing event data}

\item{eventcolor_column}{name of the event color column}

\item{fontcolor_column}{name of the lable color column}
}
\value{
same data frame as input, but with columns \code{col} and \code{fontcol} filled with color codes or names.
}
\description{
Set column col for event colors and column fontcol for label colors
}
