% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistime-package.R
\docType{package}
\name{vistime-package}
\alias{vistime-package}
\title{vistime}
\description{
Pretty timelines.
}
\author{
Sandro Raabe \email{sa.ra.online@posteo.de}
}
\keyword{gantt}
\keyword{ggplot2}
\keyword{plotly}
\keyword{posix}
\keyword{timeline}
\keyword{timelines}
\keyword{vistime}
