% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_glued.R
\name{plot_glued}
\alias{plot_glued}
\title{PLot ranges and events together as one plot}
\usage{
plot_glued(data, title, ranges, events)
}
\arguments{
\item{data}{the data frame to plot}

\item{title}{title of the plot, can be NULL}

\item{ranges}{ranges as generated by plot_ranges}

\item{events}{events as generated by plot_ranges}
}
\value{
plotly plot
}
\description{
PLot ranges and events together as one plot
}
