% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vistime-package.R
\docType{package}
\name{vistime-package}
\alias{vistime-package}
\title{vistime: Pretty Timeline Charts in R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A library for creating time based charts, like Gantt or timelines. Possible outputs
  include 'ggplot' diagrams, 'Plotly' graphs, 'Highchart' widgets and 'data.frames'. Results can be
  used in the 'RStudio' viewer pane, in 'RMarkdown' documents or in 'Shiny' apps. In the
  interactive outputs created by 'vistime()' and 'hc_vistime()', you can interact with the
  plot using mouse hover or zoom.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://shosaco.github.io/vistime/}
  \item Report bugs at \url{https://github.com/shosaco/vistime/issues}
}

}
\author{
\strong{Maintainer}: Sandro Raabe \email{sa.ra.online@posteo.de}

}
\keyword{internal}
